pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "naijaeval"
version = "0.1.0"
description = "Evaluation toolkit for AI systems in African language contexts — code-switching, dialectal robustness, and low-resource NLP."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [
    { name = "Uche Buzugbe", email = "uche.buzugbe@analyticsintelligence.com" }
]
keywords = [
    "nlp",
    "evaluation",
    "african-languages",
    "code-switching",
    "multilingual",
    "low-resource",
    "asr",
    "machine-translation",
    "benchmarks",
    "nigerian-english",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "sacrebleu>=2.3.0",
    "jiwer>=3.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tabulate>=0.9.0",
    "datasets>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]
comet = [
    "unbabel-comet>=2.2.0",
]
bertscore = [
    "bert-score>=0.3.13",
]

[project.scripts]
naijaeval = "naijaeval.cli.main:app"

[project.urls]
Homepage = "https://github.com/ubuzugbe/naijaeval"
Documentation = "https://naijaeval.readthedocs.io"
Repository = "https://github.com/ubuzugbe/naijaeval"
"Bug Tracker" = "https://github.com/ubuzugbe/naijaeval/issues"

[tool.hatch.build.targets.wheel]
packages = ["naijaeval"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"tests/**" = ["S101"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["naijaeval"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
